Update AwardPointsViewController.swift
Currently, the awarded points "Earned Points" label located in the "Your Progress" view (the Award Points View Controller) counts down to zero when a Daily Challenge is completed.  The countdown happens quite quickly and can be a little bit disconcerting if your attention is drawn away from the app while the total score is increased by the earned points leaving the Earned value at zero.  Switching the values in the call to CountingLabel.count(start: , end: ) where 0 and pointsToAward are the values for the start and end arguments respectively, allows for the same result in the total.  Additionally, it resolves the situation where the learner is distracted from the "Your Progress" view and wishes to view the "Earned" points value at a later point.